Add edit and delete functionalities for users

+++ app/Http/Controllers/Api/UserController.php
@@ -91,23 +91,74 @@ class UserController extends Controller
         $user_id = Auth::user()->id;
         $user = User::find($user_id);
 
+        if ($user_id != $user->id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+
         $validator = Validator::make($request->all(), [
-            'name' => 'required|string|max:255',
-            'description' => 'text',
-            'avatar' => 'image|mimes:png,jpg,jpeg|max:2048',
+            'name' => 'sometimes|string|max:255',
+            'description' => 'sometimes|min: 6',
+            'avatar' => 'sometimes|image|mimes:png,jpg,jpeg|max:2048',
         ]);
 
         if ($validator->fails()) {
             return response()->json($validator->errors());
         }
+
+        try {
+            if ($request->has('name')) {
+                $user->name = $request->input('name');
+            }
+            if ($request->has('description')) {
+                $user->description = $request->input('description');
+            }
+    
+            if ($request->hasFile('avatar')) {
+                $cloudinaryImage = $request->file('avatar')->storeOnCloudinary('user-profile');
+                $user->avatar = $cloudinaryImage->getSecurePath();
+               // $article->avatar_public_id = $cloudinaryImage->getPublicId();
+            }
+    
+            // Save the article
+            $user->save();
+
+            return response()->json([
+                'message' => 'Your profile has been successfully updated !',
+            ], 200);
+
+        } catch (\Exception $e) {
+            return response()->json(['message' => 'Error in updating your profile : ' . $e->getMessage()], 500);
+        }
     }
 
-    public function deleteAccount($id) {
+    public function deleteAccount() {
         $user_id = Auth::user()->id;
         $user = User::find($user_id);
+        try {
+            if ($user->articles()->exists()) {
+                // Retrieve user's articles
+                $articles = $user->articles;
+    
+                foreach ($articles as $article) {
+                    // Delete associated images and categories
+                    $article->categories()->detach();
+                    $article->images()->delete();
+                    // Delete article
+                    $article->delete();
+                }
+            }
+    
+            // Delete user account
+            $user->delete();
+    
+            return response()->json([
+                'message' => 'Your account has been successfully deleted !',
+            ], 200);
 
-       /*  $article = Article::findOrFail($id);
-        $article->delete();
-        $article->categories()->detach(); */
+        } catch (\Exception $e) {
+            return response()->json([
+                'message' => 'Error in deleting your account: ' . $e->getMessage(),
+            ], 500);
+        }
     }
 }
